lm32: Fix task scheduling
This patch fixes the issue that every 10th or so process becomes a zombie.

Currently whenever a timer irq happens task are rescheduled. There is one
problem with this approach though. When a task calls exit() it is marked as
EXIT_DEAD. The kernel will remove such a task then if all its references have
been drop, but it will not consider it for rescheduling.

Now if the timer interrupt interrupts the exit() syscall before the reference
has been dropped but after the task is marked as dead the task will never loose
all its references and thus becomes a zombie.

The solution to this problem is to never interrupt a task which is currently
running in kernel space. For this we introduce a new global variable
kernel_mode, which will be set to "1" whenever a process enters kernel mode, for
example though a syscall or when it gets interrupted by a interrupt.
Now, when the interrupt handler is entered we check kernel_mode to see if we are
allready in kernel_mode and store the result into the irq stack frame of the
current process. When the irq handler is left again we only reschedule tasks if
the current task was in user mode at the time of interruption.

This is the expected behaviour for a kernel build without CONFIG_PREEMPT.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>