lm32: Simplify mode switching
Whenever we enter kernel mode we'll switch to the kernel stack. We will always
start at the bottom of the kernel stack, once we leave kernel mode we'll be at
the bottom of the kernel stack again. Since we can calculate the kernel stack
address based on the current thread_info address (current thread_info address is
always at the lowest address of the kernel stack) we do not have to track the
kernel stack address of a process separately. Also the which_stack field is
redundant since we are always on kernel stack in kernel mode and always on user
stack in user mode, so we can remove it altogether as well.

Finally as a minor optimization put all the global variables used during mode
switch in a common struct. This is for one quite cache friendly and we only have
to load the address of the struct and can use relative addressing to access the
members instead of loading the address of each global variable individually.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>