lm32: Save the process state in the thread_info struct
Put the process state which is saved during process switching in the thread_info
stack instead of on top of the stack. This has the advantage that we know where
the registers are saved and don't need to track this, so we can finally get rid
of the ksp field of the thread_struct struct. Also only save those registers
which are callee saved. All other register will already be saved on previous
stack frames. As a result copy_thread() also looks much nicer.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>