lm32: mm: Clean up mem_init
The global num_physpages has gone away in recent times. Try to clean up
the lm32 function in a similar fashion to the other architectures.

It's strange how they've moved to use free_initmem_default, but still
have their own free_init_pages for freeing the INITRD memory. This is an
opportunity for cleanup I think.

Signed-off-by: Joel Stanley <joel@jms.id.au>